// components/ExpandableTable/types.ts
// WARNING: This file is automatically generated and managed by Superblocks CLI and should not be edited manually.
// Use the `superblocks components watch` command to regenerate this file from its source in config.ts

// All properties of your component are defined here.
// These are the properties which are surfaced in the Superblocks properties panel
// and can be referenced throughout your Superblocks Application
export interface Props {
  furnitureCategoryData: any;
  furnitureItemsData: any;
  primaryRowClicked_data: any;
  detailRowClicked_data: any;
  primaryCtaClickedType: string;
  detailCtaClickedType: string;
  latestClickType: string;
}

export interface EventTriggers {
  // Call the subsequent function(s) to trigger event(s) in Superblocks from your component.
  // These events can be wired up to event handlers in your Superblocks App
  onPrimaryCTAClick: () => void;
  onDetailRowCTAClick: () => void;
}